creada la metadata titulo y descripcion para todas las paginas.
